separate files for EntityComponents and EntityProperties classes
